refactor(toolconfigure): extract the tutorial and netrecord packages — 14 files to 8, out of the size baseline
Thirteen of the fourteen non-test files here were created by one commit
(822ef7af), so the tutorial_* and network_recording_* prefixes are that
commit's opinion, not evidence. Since then only capabilities.go and helpers.go
have been touched individually, so co-change says nothing either. The
reference graph decides, and it draws exactly two closed clusters:

  tutorial.go -> TutorialSnippets (tutorial_snippets.go)
              -> TutorialSafeAutomationLoop, TutorialCSPFallbackPlaybook
                 (tutorial_playbooks.go)
  network_recording.go -> state.TryStart/Stop/Info (network_recording_state.go)
                       -> CollectRecordedRequests (network_recording_filters.go)

Both satellite files have exactly one in-package caller and no other. Nothing
else in the package references either cluster, and the clusters do not
reference each other. The five remaining handlers (noise, capabilities,
security mode, telemetry, jitter) have zero intra-package edges at all — each
is an independent Deps-taking entry point, so they stay put. I did not merge
the three small mode-toggle handlers into one file: they share a shape but no
reference, and the package is already under the limit without it.

Deps follows the cut. tutorial gets its own three-method interface (tracking
status, pilot status, extension connectivity) and netrecord keeps the
one-method NetworkBodyProvider it already had; toolconfigure.Deps sheds the
three methods nothing in it uses any more. *ToolHandler satisfies all three,
now asserted at compile time in tools_interface_check_test.go.

Every exported symbol keeps its exact name — tutorial.TutorialSnippets stutters,
but preserving names makes the before/after API diff mechanically checkable,
which matters more here than cosmetics.

Tests move with their code, each new package getting its own narrow fake
instead of the fourteen-method fakeConfigureDeps.

Verified against a pristine origin/UNSTABLE checkout:
- every configure and generate artifact renders byte-identically from fixed
  fixtures (same harness as the previous commit, extended to configure);
- the exported declaration set across all six packages is unchanged except the
  two new narrow Deps interfaces;
- the multiset of non-comment production code lines is unchanged except the
  two deleted helper wrappers and those interfaces.
