+++ cmd/browser-agent/tools_async_result_enrichment_csp.go
@@ -7,7 +7,7 @@ import (
 	"encoding/json"
 	"strings"
 
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/toolconfigure"
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/toolconfigure/tutorial"
 )
 
 func annotateCSPFailure(responseData map[string]any, cmdError string, result json.RawMessage) {
@@ -28,7 +28,7 @@ func annotateCSPFailure(responseData map[string]any, cmdError string, result jso
 		}
 	}
 	if _, exists := responseData["retry"]; !exists {
-		responseData["retry"] = toolconfigure.CSPRetryNavigationGuidance
+		responseData["retry"] = tutorial.CSPRetryNavigationGuidance
 	}
 }
