+++ cmd/browser-agent/tools_configure_network_recording.go
@@ -1,4 +1,4 @@
-// Purpose: Delegates network recording handler to the toolconfigure sub-package.
+// Purpose: Delegates network recording handler to the toolconfigure/netrecord sub-package.
 // Why: Keeps the handler wiring in main thin while logic lives in the sub-package.
 // Docs: docs/features/feature/backend-log-streaming/index.md
 
@@ -7,10 +7,10 @@ package main
 import (
 	"encoding/json"
 
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/toolconfigure"
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/cmd/browser-agent/internal/toolconfigure/netrecord"
 )
 
 // toolConfigureNetworkRecording delegates to the sub-package handler.
 func (h *ToolHandler) toolConfigureNetworkRecording(req JSONRPCRequest, args json.RawMessage) JSONRPCResponse {
-	return toolconfigure.HandleNetworkRecording(h.capture, h.networkRecording, req, args)
+	return netrecord.HandleNetworkRecording(h.capture, h.networkRecording, req, args)
 }
